Time: 92 ms (29.67%), Space: 169.3 MB (10.53%) - LeetHub

+++ 3653-xor-after-range-multiplication-queries-i/3653-xor-after-range-multiplication-queries-i.cpp
@@ -8,7 +8,7 @@ public:
                 nums[i]=((nums[i]*1LL)*v)%mod;
             }
         }
-        long long ans=nums[0];
+        int ans=nums[0];
         for(int i=1;i<nums.size();i++){
             ans^=nums[i];
         }
